requirements.txt
pycryptodome
pytest
Pillow